+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add method explain for explain query to QuerySession and QuerySessionPool classes
+
 ## 3.21.8 ##
 * Fix: convert gRPC stream termination to YDB errors in async query client